+++ src/main/java/com/jaeyeonling/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package com.jaeyeonling.racingcar;
+package com.jaeyeonling.racingcar.domain;
 
 class Car {
