+++ src/main/java/com/jaeyeonling/racingcar/domain/MoveStrategy.java
@@ -1,4 +1,4 @@
-package com.jaeyeonling.racingcar;
+package com.jaeyeonling.racingcar.domain;
 
 public interface MoveStrategy {
