+++ src/main/java/com/jaeyeonling/racingcar/domain/RacingGameStatus.java
@@ -1,4 +1,4 @@
-package com.jaeyeonling.racingcar;
+package com.jaeyeonling.racingcar.domain;
 
 public class RacingGameStatus implements Visualizable {
