+++ src/main/java/com/jaeyeonling/racingcar/domain/Visualizable.java
@@ -1,4 +1,4 @@
-package com.jaeyeonling.racingcar;
+package com.jaeyeonling.racingcar.domain;
 
 public interface Visualizable {
