+++ src/test/java/com/jaeyeonling/racingcar/CarTest.java
@@ -1,7 +1,7 @@
 package com.jaeyeonling.racingcar;
 
+import com.jaeyeonling.racingcar.domain.Car;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
